refactor(pm): fold binary_resolver into the execute service
find_binary / find_binary_in_cache / find_binary_in_hierarchy were the
binary-resolution mechanism of `utoo x`, but lived in util/ as a generic
"utility" with a pub surface — execute.rs was their only consumer. They
are execute-domain logic, not reusable helpers, so move them (and their
tests) into service/execute.rs as private fns and delete
util/binary_resolver.rs. No behavior change; the pub indirection is gone.
