+++ crates/pm/src/service/execute.rs
@@ -1,15 +1,17 @@
-use crate::service::package_management::PackageManagementService;
-use crate::util::binary_resolver;
-use anyhow::{Result, anyhow};
-use std::path::Path;
+use std::fs;
+use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
+
+use anyhow::{Result, anyhow};
+
+use crate::service::package_management::PackageManagementService;
 
 /// Execute a package binary
 pub async fn execute_package(command: &str, args: Vec<String>) -> Result<()> {
     tracing::debug!("Executing command: {command} with args: {args:?}");
 
     // First, try to find the binary in local node_modules/.bin directories
-    if let Some(binary_path) = binary_resolver::find_binary(command).await? {
+    if let Some(binary_path) = find_binary(command).await? {
         tracing::debug!("Found binary at: {}", binary_path.display());
         return execute_binary(&binary_path, args).await;
     }
@@ -30,7 +32,7 @@ pub async fn execute_package(command: &str, args: Vec<String>) -> Result<()> {
     // utoo -x eslint --version
     // utoo -x @modelcontextprotocol/create-server --version
     // utoo -x @modelcontextprotocol/create-server create-mcp-server --version
-    match binary_resolver::find_binary_in_cache(&package_cache_dir).await {
+    match find_binary_in_cache(&package_cache_dir).await {
         Ok(Some(binary_path)) => {
             tracing::debug!("Found binary in cache at: {}", binary_path.display());
             execute_binary(&binary_path, args).await
@@ -82,5 +84,139 @@ async fn execute_binary(binary_path: &Path, args: Vec<String>) -> Result<()> {
     } else {
         let exit_code = status.code().unwrap_or(-1);
         Err(anyhow!("Command failed with exit code: {exit_code}"))
+    }
+}
+
+/// Find `command` in `node_modules/.bin`, searching up from the cwd — the
+/// standard mechanism for running a locally-installed tool's bin shim.
+async fn find_binary(command: &str) -> Result<Option<PathBuf>> {
+    let current_dir = std::env::current_dir()?;
+    find_binary_in_hierarchy(&current_dir, command).await
+}
+
+/// Find the binary of a tool freshly installed into the utx cache: return the
+/// first file in `<package_cache_dir>/bin`.
+async fn find_binary_in_cache(package_cache_dir: &Path) -> Result<Option<PathBuf>> {
+    let bin_dir = package_cache_dir.join("bin");
+
+    if !crate::fs::try_exists(&bin_dir).await? {
+        return Ok(None);
+    }
+
+    // Get the first file in bin directory
+    let entries = fs::read_dir(&bin_dir)?;
+
+    for entry in entries {
+        let entry = entry?;
+        let path: PathBuf = entry.path();
+        if path.is_file() {
+            return Ok(Some(path));
+        }
+    }
+
+    Ok(None)
+}
+
+/// Search for binary in the directory hierarchy starting from the given path.
+async fn find_binary_in_hierarchy(start_path: &Path, command: &str) -> Result<Option<PathBuf>> {
+    let mut current_path = start_path.to_path_buf();
+
+    loop {
+        let bin_path = current_path.join("node_modules").join(".bin").join(command);
+
+        // Check if the binary exists and is executable
+        if crate::fs::try_exists(&bin_path).await? {
+            return Ok(Some(bin_path));
+        }
+
+        // Move up to parent directory
+        if let Some(parent) = current_path.parent() {
+            current_path = parent.to_path_buf();
+        } else {
+            // Reached the root directory
+            break;
+        }
+    }
+
+    Ok(None)
+}
+
+#[cfg(test)]
+mod tests {
+    #[cfg(unix)]
+    use std::os::unix::fs::PermissionsExt;
+
+    use tempfile::TempDir;
+
+    use super::*;
+    use crate::fs as async_fs;
+
+    // Helper function to create a test directory structure
+    async fn create_test_structure() -> Result<TempDir> {
+        let temp_dir = TempDir::new()?;
+        let base_path = temp_dir.path();
+
+        // Create node_modules/.bin directory
+        let bin_dir = base_path.join("node_modules").join(".bin");
+        async_fs::create_dir_all(&bin_dir).await?;
+
+        // Create a test binary file
+        let test_binary = bin_dir.join("test-cmd");
+        async_fs::write(&test_binary, "#!/bin/bash\necho 'test'").await?;
+
+        // Set executable permissions on Unix systems
+        #[cfg(unix)]
+        {
+            let mut perms = async_fs::metadata(&test_binary).await?.permissions();
+            perms.set_mode(0o755);
+            async_fs::set_permissions(&test_binary, perms).await?;
+        }
+
+        Ok(temp_dir)
+    }
+
+    #[tokio::test]
+    async fn test_find_binary_in_hierarchy_found() {
+        let temp_dir = create_test_structure().await.unwrap();
+        let start_path = temp_dir.path();
+
+        let result = find_binary_in_hierarchy(start_path, "test-cmd")
+            .await
+            .unwrap();
+        assert!(result.is_some());
+
+        let found_path = result.unwrap();
+        assert!(found_path.ends_with("node_modules/.bin/test-cmd"));
+        assert!(found_path.exists());
+    }
+
+    #[tokio::test]
+    async fn test_find_binary_in_hierarchy_not_found() {
+        let temp_dir = create_test_structure().await.unwrap();
+        let start_path = temp_dir.path();
+
+        let result = find_binary_in_hierarchy(start_path, "nonexistent-cmd")
+            .await
+            .unwrap();
+        assert!(result.is_none());
+    }
+
+    #[tokio::test]
+    async fn test_find_binary_in_hierarchy_search_up() {
+        let temp_dir = create_test_structure().await.unwrap();
+        let base_path = temp_dir.path();
+
+        // Create a subdirectory without node_modules
+        let sub_dir = base_path.join("subdir").join("deeper");
+        async_fs::create_dir_all(&sub_dir).await.unwrap();
+
+        // Search from the subdirectory should find the binary in parent
+        let result = find_binary_in_hierarchy(&sub_dir, "test-cmd")
+            .await
+            .unwrap();
+        assert!(result.is_some());
+
+        let found_path = result.unwrap();
+        assert!(found_path.ends_with("node_modules/.bin/test-cmd"));
     }
 }
